Deduplicate shared tool and fuzzy-match helpers
